refactor: move client to internal/client
